requirements.txt
django==3.2.12
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
psycopg2==2.9.3
requests==2.27.1
flake8==4.0.1
